Move TransactionStatistics into StorageEngine
