+++ arangod/StorageEngine/StorageEngine.h
@@ -27,8 +27,8 @@
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Basics/Result.h"
 #include "Indexes/IndexFactory.h"
-#include "Statistics/ServerStatistics.h"
 #include "StorageEngine/HealthData.h"
+#include "StorageEngine/TransactionStatistics.h"
 #include "Transaction/ManagerFeature.h"
 #include "Transaction/OperationOrigin.h"
 #include "VocBase/Identifiers/DataSourceId.h"
